refactor: 에러 관리 로직 커스텀 훅으로 분리
